refac: move static to utils
