+++ src/utils/static.ts
@@ -23,3 +23,10 @@ export const CONTRIBUTION_BOX_COLOR = "#646E7A";
 export const CONTRIBUTION_FONT_COLOR = "#CFDAE4";
 
 export const CONTRIBUTION_FONT_SIZE = 12;
+
+export const DefaultConfigValue = {
+  color: "GREEN" as const,
+  grassSpan: 30,
+  month: 3,
+  darkmode: false,
+};
